requirements.txt
astroid==2.4.2
colorama==0.4.4
isort==5.7.0
lazy-object-proxy==1.4.3
lxml==4.6.3
mccabe==0.6.1
numpy==1.20.1
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
Pillow==8.2.0
pylint==2.6.2
pytesseract==0.3.7
python-docx==0.8.10
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
